refactor(requirements-manager): split requirements-checker.utils.ts by category
The 1571-line requirements-checker.utils.ts was a router + retry harness +
all 25 check function implementations in one file. Split the check
implementations into category modules under src/requirements-manager/checks/
so each lives next to its peers and the router stays small.

  - checks/grafana-api.ts (505 lines): permissions, roles, data sources,
    plugins, dashboards, login, admin/editor — anything that hits a
    Grafana API or reads from config.bootData.
  - checks/location.ts (33 lines): on-page check (returns location fix
    metadata).
  - checks/env.ts (134 lines): feature toggles, build environment,
    minimum version, renderer context.
  - checks/vars.ts (97 lines): guide-response variables (var-X:Y).
  - checks/terminal.ts (31 lines): is-terminal-active (dynamic import
    of the coda integration).

requirements-checker.utils.ts now holds only:
  - The route map (routeUnifiedCheck)
  - The retry harness (executeChecksWithRetry)
  - The exported entry points (checkRequirements, checkPostconditions)
  - validateInteractiveRequirements (author-time validation)

Net effect: the router file shrinks 1571 → 386 lines (-75%); adding a new
requirement type is now an edit in one small file plus a router case
rather than a scroll through 1500 lines of mixed concerns.

Architecture ratchet entries updated to point at the new file paths
(both pre-existing violations — `requirements-manager -> integrations`
for the dynamic terminal import and `requirements-manager -> context-engine`
for ContextService — moved with the code, no net new violations).
