+++ src/requirements-manager/requirements-checker.utils.ts
@@ -1,19 +1,43 @@
 /**
- * Pure requirements checking utilities
- * Extracted from interactive.hook.ts to eliminate mock element anti-pattern
+ * Requirements checking — router and retry harness.
  *
- * This module handles requirements checking without DOM manipulation,
- * focusing on API calls, configuration checks, and Grafana state validation.
+ * The individual check implementations live next to their peers in
+ * `./checks/*` (grafana-api, location, env, vars, terminal). This file holds
+ * only:
+ *   - The route map (which check string maps to which function)
+ *   - The retry harness used by `checkRequirements` / `checkPostconditions`
+ *   - The exported entry points
+ *   - Author-time validation (`validateInteractiveRequirements`)
+ *
+ * Adding a new requirement type means:
+ *   1. Implement the check in the right `./checks/*` file (or a new one)
+ *   2. Add a case to `routeUnifiedCheck`
+ *   3. Add the requirement string to `isValidRequirement` in `types/requirements.types.ts`
  */
 
-import { locationService, config, hasPermission, getDataSourceSrv, getBackendSrv } from '@grafana/runtime';
-// eslint-disable-next-line no-restricted-imports -- [ratchet] ALLOWED_LATERAL_VIOLATIONS: requirements-manager -> context-engine
-import { ContextService } from '../context-engine';
 import { reftargetExistsCheck, navmenuOpenCheck, sectionCompletedCheck, formValidCheck } from '../lib/dom';
 import { isValidRequirement } from '../types/requirements.types';
 import { INTERACTIVE_CONFIG } from '../constants/interactive-config';
 import { TimeoutManager } from '../utils/timeout-manager';
-import { guideResponseStorage } from '../lib/user-storage';
+
+import {
+  hasPermissionCheck,
+  hasRoleCheck,
+  hasDataSourceCheck,
+  hasPluginCheck,
+  hasDashboardNamedCheck,
+  isAdminCheck,
+  isLoggedInCheck,
+  isEditorCheck,
+  hasDatasourcesCheck,
+  pluginEnabledCheck,
+  dashboardExistsCheck,
+  datasourceConfiguredCheck,
+} from './checks/grafana-api';
+import { onPageCheck } from './checks/location';
+import { hasFeatureCheck, inEnvironmentCheck, minVersionCheck, rendererCheck } from './checks/env';
+import { guideVariableCheck } from './checks/vars';
+import { terminalActiveCheck } from './checks/terminal';
 
 // Re-export types for convenience
 export interface RequirementsCheckResult {
@@ -49,10 +73,6 @@ export interface RequirementsCheckOptions {
   scrollContainer?: string;
 }
 
-/**
- * Core requirements checking function (pure implementation)
- * Replaces the mock element anti-pattern with direct string-based checking
- */
 type CheckMode = 'pre' | 'post';
 
 interface CheckContext {
@@ -203,8 +223,7 @@ async function runUnifiedChecks(
 }
 
 /**
- * Shared retry logic for requirements and postconditions checking
- * Extracts common retry pattern to reduce duplication
+ * Shared retry logic for requirements and postconditions checking.
  */
 async function executeChecksWithRetry(
   options: RequirementsCheckOptions,
@@ -301,1230 +320,26 @@ async function executeChecksWithRetry(
 }
 
 /**
- * Pre-action requirements checker
- * Validates requirements before an action can be performed
+ * Pre-action requirements checker. Validates requirements before an action can run.
  */
 export async function checkRequirements(options: RequirementsCheckOptions): Promise<RequirementsCheckResult> {
   return executeChecksWithRetry(options, 'pre', 'requirements');
 }
 
 /**
- * Post-action verification checker
- * Similar to checkRequirements, but semantically intended for verifying outcomes AFTER an action.
- * Uses the same underlying pure checks where applicable (e.g., has-plugin, has-datasource, has-dashboard-named, on-page).
- * Excludes pre-action gating like navmenu-open and existence checks that are about enabling interactions.
+ * Post-action verification checker. Same underlying checks as `checkRequirements`,
+ * intended for verifying outcomes AFTER an action.
  */
 export async function checkPostconditions(options: RequirementsCheckOptions): Promise<RequirementsCheckResult> {
   return executeChecksWithRetry(options, 'post', 'postconditions');
 }
 
 /**
- * ============================================================================
- * PURE REQUIREMENTS CHECKING FUNCTIONS
- * These functions only use APIs, configuration, and Grafana state - no DOM
- * ============================================================================
- */
-
-/**
- * Permission checking - verifies user has specific Grafana permissions
- *
- * Use cases:
- * - Admin features: ensure user can manage data sources, users, etc.
- * - Data access: verify read/write permissions for specific resources
- * - Security gates: prevent unauthorized access to sensitive operations
- * - Role-based tutorials: show different content based on permissions
- *
- * How it works:
- * - Uses Grafana's built-in hasPermission() function
- * - Checks against Grafana's permission system
- * - Returns immediate boolean result (no API calls)
- *
- * Example usage:
- * - data-requirements="has-permission:datasources.read" - can view data sources
- * - data-requirements="has-permission:users.write" - can manage users
- * - data-requirements="has-permission:dashboards.create" - can create dashboards
- *
- * Common permissions:
- * - datasources.read, datasources.write
- * - dashboards.read, dashboards.write, dashboards.create
- * - users.read, users.write
- * - orgs.read, orgs.write
- */
-async function hasPermissionCheck(check: string): Promise<CheckResultError> {
-  try {
-    const permission = check.replace('has-permission:', '');
-    const hasAccess = hasPermission(permission);
-
-    return {
-      requirement: check,
-      pass: hasAccess,
-      error: hasAccess ? undefined : `Missing permission: ${permission}`,
-      context: { permission, hasAccess },
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Permission check failed: ${error}`,
-      context: { error: String(error) },
-    };
-  }
-}
-
-/**
- * User role checking - verifies user has specific organizational role
- *
- * Use cases:
- * - Role-based access: show admin-only or editor-only features
- * - Tutorial branching: different paths for different user types
- * - Feature gating: ensure user has sufficient privileges
- * - Security validation: verify role before sensitive operations
- *
- * Supported roles:
- * - admin/grafana-admin: Grafana admin or org admin
- * - editor: Editor role or higher (includes admins)
- * - viewer: Any role (viewer, editor, admin)
- * - Custom roles: exact case-insensitive match
- *
- * Role hierarchy (higher roles include lower permissions):
- * - Grafana Admin > Org Admin > Editor > Viewer
- *
- * Example usage:
- * - data-requirements="has-role:admin" - admin or grafana admin only
- * - data-requirements="has-role:editor" - editor or admin
- * - data-requirements="has-role:viewer" - any authenticated user
- * - data-requirements="has-role:custom-role" - exact role match
- */
-async function hasRoleCheck(check: string): Promise<CheckResultError> {
-  try {
-    const user = config.bootData?.user;
-    if (!user) {
-      return {
-        requirement: check,
-        pass: false,
-        error: 'User information not available',
-        context: null,
-      };
-    }
-
-    const requiredRole = check.replace('has-role:', '').toLowerCase();
-    let hasRole = false;
-
-    switch (requiredRole) {
-      case 'admin':
-      case 'grafana-admin':
-        hasRole = user.isGrafanaAdmin === true || user.orgRole === 'Admin';
-        break;
-      case 'editor':
-        hasRole = user.orgRole === 'Editor' || user.orgRole === 'Admin' || user.isGrafanaAdmin === true;
-        break;
-      case 'viewer':
-        hasRole = !!user.orgRole; // Any role satisfies viewer requirement
-        break;
-      default:
-        // For custom roles, do case-insensitive comparison
-        hasRole = user.orgRole?.toLowerCase() === requiredRole;
-    }
-
-    return {
-      requirement: check,
-      pass: hasRole,
-      error: hasRole
-        ? undefined
-        : `User role '${user.orgRole || 'none'}' does not meet requirement '${requiredRole}' (isGrafanaAdmin: ${user.isGrafanaAdmin})`,
-      context: {
-        orgRole: user.orgRole,
-        isGrafanaAdmin: user.isGrafanaAdmin,
-        requiredRole,
-        userId: user.id,
-      },
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Role check failed: ${error}`,
-      context: { error: String(error) },
-    };
-  }
-}
-
-/**
- * Data source existence checking - verifies a specific data source is installed
- *
- * Use cases:
- * - Tutorial prerequisites: ensure required data source is available
- * - Feature dependencies: check if specific data source type exists
- * - Multi-data source tutorials: verify all needed sources are present
- * - Troubleshooting: validate data source installation
- *
- * How it works:
- * - Searches by name or type (case-insensitive)
- * - Uses DataSourceSrv.getList() for immediate results
- * - No connection testing (use datasource-configured for that)
- *
- * Example usage:
- * - data-requirements="has-datasource:prometheus" - Prometheus type exists
- * - data-requirements="has-datasource:loki" - Loki type exists
- * - data-requirements="has-datasource:My Production DB" - exact name match
- *
- * Search priority:
- * 1. Exact name match
- * 2. Exact type match
- *
- * Difference from datasource-configured:
- * - has-datasource: checks if data source exists/is installed
- * - datasource-configured: checks if data source exists AND works
- */
-async function hasDataSourceCheck(check: string): Promise<CheckResultError> {
-  try {
-    const dataSourceSrv = getDataSourceSrv();
-    const dsRequirement = check.replace('has-datasource:', '').toLowerCase();
-
-    const dataSources = dataSourceSrv.getList();
-    let found = false;
-    let matchType = '';
-
-    // Check for exact matches in name or type, then normalized type
-    // Type normalization strips common prefixes/suffixes (e.g. grafana-testdata-datasource → testdata)
-    for (const ds of dataSources) {
-      if (ds.name.toLowerCase() === dsRequirement) {
-        found = true;
-        matchType = 'name';
-        break;
-      }
-      if (ds.type.toLowerCase() === dsRequirement) {
-        found = true;
-        matchType = 'type';
-        break;
-      }
-      const normalizedType = ds.type
-        .toLowerCase()
-        .replace(/^grafana-/, '')
-        .replace(/-datasource$/, '');
-      if (normalizedType === dsRequirement) {
-        found = true;
-        matchType = 'type-normalized';
-        break;
-      }
-    }
-
-    return {
-      requirement: check,
-      pass: found,
-      error: found ? undefined : `No data source found with name/type: ${dsRequirement}`,
-      context: {
-        searched: dsRequirement,
-        matchType: found ? matchType : null,
-        available: dataSources.map((ds) => ({ name: ds.name, type: ds.type, uid: ds.uid })),
-      },
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Data source check failed: ${error}`,
-      context: { error },
-    };
-  }
-}
-
-/**
- * Plugin installation checking - verifies a specific plugin is installed
- *
- * Use cases:
- * - Tutorial prerequisites: ensure required plugins are available
- * - Feature availability: check if optional plugins are installed
- * - Compatibility checking: verify plugin dependencies
- * - Installation guides: validate plugin setup
- *
- * How it works:
- * - Searches by exact plugin ID match
- * - Uses /api/plugins endpoint via ContextService
- * - Only checks installation, not enabled status
- *
- * Example usage:
- * - data-requirements="has-plugin:volkovlabs-rss-datasource" - RSS plugin installed
- * - data-requirements="has-plugin:grafana-clock-panel" - Clock panel installed
- * - data-requirements="has-plugin:alexanderzobnin-zabbix-app" - Zabbix app installed
- *
- * Plugin ID format:
- * - Use exact plugin ID from plugin.json
- * - Case-sensitive matching
- * - Check Grafana plugin catalog for correct IDs
- *
- * Difference from plugin-enabled:
- * - has-plugin: checks if plugin is installed (may be disabled)
- * - plugin-enabled: checks if plugin is installed AND enabled
- */
-async function hasPluginCheck(check: string): Promise<CheckResultError> {
-  try {
-    const pluginId = check.replace('has-plugin:', '');
-    const plugins = await ContextService.fetchPlugins();
-    const pluginExists = plugins.some((plugin) => plugin.id === pluginId);
-
-    return {
-      requirement: check,
-      pass: pluginExists,
-      error: pluginExists ? undefined : `Plugin '${pluginId}' is not installed or enabled`,
-      context: {
-        searched: pluginId,
-        totalPlugins: plugins.length,
-        // More actionable: tell them how to find what they need
-        suggestion:
-          plugins.length > 0
-            ? `Check your Grafana plugin management page - ${plugins.length} plugins are available`
-            : 'No plugins found - check your Grafana installation',
-      },
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Plugin check failed: ${error}`,
-      context: { error },
-    };
-  }
-}
-
-/**
- * Dashboard existence by name - verifies a specific dashboard exists by title
- *
- * Use cases:
- * - Tutorial dependencies: ensure required dashboards are available
- * - Dashboard-based tutorials: verify target dashboard exists before proceeding
- * - Data visualization guides: check if example dashboards are present
- * - Import validation: confirm dashboard was successfully imported
- *
- * How it works:
- * - Searches dashboards by title using /api/search endpoint
- * - Case-insensitive exact title matching
- * - Returns search results for debugging
- *
- * Example usage:
- * - data-requirements="has-dashboard-named:Node Exporter Full" - specific dashboard
- * - data-requirements="has-dashboard-named:My Custom Dashboard" - user-created dashboard
- * - data-requirements="has-dashboard-named:Prometheus Overview" - imported dashboard
- *
- * Search behavior:
- * - Exact title match required (case-insensitive)
- * - Partial matches are found but don't satisfy requirement
- * - Provides helpful suggestions when similar dashboards exist
- *
- * Difference from dashboard-exists:
- * - has-dashboard-named: checks for specific dashboard by title
- * - dashboard-exists: checks if any dashboards exist in the system
- */
-async function hasDashboardNamedCheck(check: string): Promise<CheckResultError> {
-  try {
-    const dashboardName = check.replace('has-dashboard-named:', '');
-    const dashboards = await ContextService.fetchDashboardsByName(dashboardName);
-    const dashboardExists = dashboards.some(
-      (dashboard) => dashboard.title.toLowerCase() === dashboardName.toLowerCase()
-    );
-
-    return {
-      requirement: check,
-      pass: dashboardExists,
-      error: dashboardExists ? undefined : `Dashboard named '${dashboardName}' not found`,
-      context: {
-        searched: dashboardName,
-        totalFound: dashboards.length,
-        suggestion:
-          dashboards.length > 0
-            ? `Found ${dashboards.length} dashboards matching search, but none with exact name '${dashboardName}'. Check dashboard names in Grafana.`
-            : `No dashboards found matching '${dashboardName}'. Check if the dashboard exists.`,
-      },
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Dashboard check failed: ${error}`,
-      context: { error },
-    };
-  }
-}
-
-/**
- * Page location checking - verifies user is on a specific page/URL
- *
- * Use cases:
- * - Context-sensitive tutorials: show steps only on relevant pages
- * - Navigation validation: ensure user reached the correct page
- * - Page-specific features: gate functionality to appropriate locations
- * - Workflow enforcement: verify user followed navigation steps
- *
- * How it works:
- * - Gets current pathname from Grafana's locationService
- * - Supports both exact matches and partial matches (contains)
- * - Case-sensitive path matching
- *
- * Example usage:
- * - data-requirements="on-page:/dashboards" - on dashboards page
- * - data-requirements="on-page:/datasources" - on data sources page
- * - data-requirements="on-page:/d/" - on any dashboard view page
- * - data-requirements="on-page:/explore" - on explore page
- *
- * Common Grafana paths:
- * - /dashboards - dashboard list
- * - /datasources - data source management
- * - /d/[uid] - specific dashboard
- * - /explore - explore/query interface
- * - /alerting - alerting rules
- * - /admin - admin pages
- *
- * Matching behavior:
- * - Exact match: current path === required path
- * - Partial match: current path contains required path
- */
-async function onPageCheck(check: string): Promise<CheckResultError> {
-  try {
-    const location = locationService.getLocation();
-    const requiredPath = check.replace('on-page:', '');
-    const currentPath = location.pathname;
-    const matches = currentPath.includes(requiredPath) || currentPath === requiredPath;
-
-    return {
-      requirement: check,
-      pass: matches,
-      error: matches ? undefined : `Current page '${currentPath}' does not match required path '${requiredPath}'`,
-      canFix: !matches,
-      fixType: matches ? undefined : 'location',
-      targetHref: matches ? undefined : requiredPath,
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Page check failed: ${error}`,
-    };
-  }
-}
-
-/**
- * Feature toggle checking - verifies a specific feature flag is enabled
- *
- * Use cases:
- * - Beta features: gate experimental functionality
- * - Version-specific features: check if new features are available
- * - A/B testing: show different content based on feature flags
- * - Gradual rollouts: enable features for specific deployments
- *
- * How it works:
- * - Checks config.featureToggles object
- * - Returns true only if feature is explicitly enabled
- * - Uses exact feature name matching
- *
- * Example usage:
- * - data-requirements="has-feature:alertingPreview" - alerting preview enabled
- * - data-requirements="has-feature:newDashboardExperience" - new dashboard UI
- * - data-requirements="has-feature:publicDashboards" - public sharing enabled
- *
- * Common feature toggles:
- * - alertingPreview - new alerting system
- * - publicDashboards - public dashboard sharing
- * - newDashboardExperience - updated dashboard UI
- * - correlations - data source correlations
- *
- * Note: Feature names are case-sensitive and match Grafana's feature toggle names
- */
-async function hasFeatureCheck(check: string): Promise<CheckResultError> {
-  try {
-    const featureName = check.replace('has-feature:', '');
-    const featureToggles = config.featureToggles as Record<string, boolean> | undefined;
-    const isEnabled = featureToggles && featureToggles[featureName];
-
-    return {
-      requirement: check,
-      pass: !!isEnabled,
-      error: isEnabled ? undefined : `Feature toggle '${featureName}' is not enabled`,
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Feature check failed: ${error}`,
-    };
-  }
-}
-
-/**
- * Environment checking - verifies Grafana is running in a specific environment
- *
- * Use cases:
- * - Cloud-specific features: show cloud-only functionality
- * - Development tutorials: different steps for dev vs production
- * - Environment-specific configuration: adapt content to deployment type
- * - Feature availability: some features only work in certain environments
- *
- * How it works:
- * - Checks config.buildInfo.env value
- * - Case-insensitive environment matching
- * - Uses Grafana's build-time environment detection
- *
- * Example usage:
- * - data-requirements="in-environment:production" - production Grafana
- * - data-requirements="in-environment:development" - dev environment
- * - data-requirements="in-environment:cloud" - Grafana Cloud
- *
- * Common environments:
- * - production - standard production deployment
- * - development - development/testing environment
- * - cloud - Grafana Cloud instances
- *
- * Note: Environment is set at build time and reflects deployment type
- */
-async function inEnvironmentCheck(check: string): Promise<CheckResultError> {
-  try {
-    const requiredEnv = check.replace('in-environment:', '').toLowerCase();
-    const currentEnv = config.buildInfo?.env?.toLowerCase() || 'unknown';
-
-    return {
-      requirement: check,
-      pass: currentEnv === requiredEnv,
-      error:
-        currentEnv === requiredEnv
-          ? undefined
-          : `Current environment '${currentEnv}' does not match required '${requiredEnv}'`,
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Environment check failed: ${error}`,
-    };
-  }
-}
-
-/**
- * Minimum version checking - verifies Grafana version meets minimum requirement
- *
- * Use cases:
- * - Feature compatibility: ensure features are available in current version
- * - API availability: check if specific APIs exist in this Grafana version
- * - Tutorial adaptation: show different steps for different versions
- * - Deprecation warnings: alert about unsupported older versions
- *
- * How it works:
- * - Compares current Grafana version with required minimum
- * - Uses semantic versioning comparison (major.minor.patch)
- * - Gets version from config.buildInfo.version
- *
- * Example usage:
- * - data-requirements="min-version:9.0.0" - requires Grafana 9.0+
- * - data-requirements="min-version:10.1.0" - requires Grafana 10.1+
- * - data-requirements="min-version:8.5.0" - requires Grafana 8.5+
- *
- * Version comparison logic:
- * - 10.1.0 meets min-version:9.0.0 ✅
- * - 9.5.2 meets min-version:9.5.0 ✅
- * - 8.4.0 fails min-version:9.0.0 ❌
- *
- * Use for:
- * - New API features introduced in specific versions
- * - UI changes that happened in certain releases
- * - Plugin compatibility requirements
- */
-async function minVersionCheck(check: string): Promise<CheckResultError> {
-  try {
-    const requiredVersion = check.replace('min-version:', '');
-    const currentVersion = config.buildInfo?.version || '0.0.0';
-
-    const parseVersion = (v: string): [number, number, number] => {
-      const parts = v.split('.').map((n) => parseInt(n, 10));
-      return [parts[0] ?? 0, parts[1] ?? 0, parts[2] ?? 0];
-    };
-    const [reqMajor, reqMinor, reqPatch] = parseVersion(requiredVersion);
-    const [curMajor, curMinor, curPatch] = parseVersion(currentVersion);
-
-    const meetsRequirement =
-      curMajor > reqMajor ||
-      (curMajor === reqMajor && curMinor > reqMinor) ||
-      (curMajor === reqMajor && curMinor === reqMinor && curPatch >= reqPatch);
-
-    return {
-      requirement: check,
-      pass: meetsRequirement,
-      error: meetsRequirement
-        ? undefined
-        : `Current version '${currentVersion}' does not meet minimum requirement '${requiredVersion}'`,
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Version check failed: ${error}`,
-    };
-  }
-}
-
-/**
- * Admin status checking - verifies user has admin privileges (shorthand for has-role:admin)
- *
- * Use cases:
- * - Admin-only features: gate administrative functionality
- * - System configuration: ensure user can modify system settings
- * - User management: verify permissions for user operations
- * - Security-sensitive operations: require admin privileges
- *
- * How it works:
- * - Delegates to hasRoleCheck('has-role:admin') for consistency
- * - Checks both isGrafanaAdmin and orgRole === 'Admin'
- * - Provides convenient shorthand for common admin check
- *
- * Example usage:
- * - data-requirements="is-admin" - simpler than has-role:admin
- * - Useful in admin tutorials and configuration guides
- *
- * Admin types recognized:
- * - Grafana Admin (isGrafanaAdmin: true) - super admin
- * - Organization Admin (orgRole: 'Admin') - org-level admin
- *
- * Equivalent to: has-role:admin
- */
-async function isAdminCheck(check: string): Promise<CheckResultError> {
-  // Just call hasRoleCheck with 'has-role:admin' to ensure identical logic
-  const result = await hasRoleCheck('has-role:admin');
-
-  // Update the requirement field to match the original check
-  return {
-    ...result,
-    requirement: check,
-  };
-}
-
-/**
- * Login status checking - verifies user is authenticated
- *
- * Use cases:
- * - Authentication gates: ensure user is logged in before proceeding
- * - User-specific features: gate personalized functionality
- * - Security validation: verify authentication for sensitive operations
- * - Tutorial prerequisites: ensure user has valid session
- *
- * How it works:
- * - Checks config.bootData.user existence and isSignedIn flag
- * - Validates both user object and sign-in status
- * - No API calls required (uses cached user data)
- *
- * Example usage:
- * - data-requirements="is-logged-in" - user must be authenticated
- * - Common prerequisite for most interactive guides
- * - Prevents anonymous users from accessing user-specific features
- *
- * Authentication states:
- * - Logged in: user exists and isSignedIn === true
- * - Not logged in: no user or isSignedIn === false
- * - Unknown: user data not available (treats as not logged in)
- *
- * Note: This is a basic authentication check, not authorization
- */
-async function isLoggedInCheck(check: string): Promise<CheckResultError> {
-  try {
-    const user = config.bootData?.user;
-    const isLoggedIn = !!user && !!user.isSignedIn;
-
-    return {
-      requirement: check,
-      pass: isLoggedIn,
-      error: isLoggedIn ? undefined : 'User is not logged in',
-      context: {
-        hasUser: !!user,
-        isSignedIn: user?.isSignedIn,
-        userId: user?.id,
-      },
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Login check failed: ${error}`,
-      context: { error },
-    };
-  }
-}
-
-/**
- * Editor role checking - verifies user has editor privileges or higher
- *
- * Use cases:
- * - Content editing: ensure user can modify dashboards, panels, etc.
- * - Configuration changes: verify permissions for system modifications
- * - Tutorial gating: show editing features only to authorized users
- * - Write operations: validate permissions before allowing changes
- *
- * How it works:
- * - Checks for Editor, Admin, or Grafana Admin roles
- * - Follows role hierarchy (admins have editor permissions)
- * - Uses cached user data from config.bootData.user
- *
- * Example usage:
- * - data-requirements="is-editor" - can edit content
- * - Useful before dashboard editing tutorials
- * - Gates features that modify Grafana configuration
- *
- * Role hierarchy (all can edit):
- * - Grafana Admin (isGrafanaAdmin: true)
- * - Organization Admin (orgRole: 'Admin')
- * - Editor (orgRole: 'Editor')
- *
- * Excluded:
- * - Viewer (orgRole: 'Viewer') - read-only access
- *
- * Note: More permissive than is-admin, includes editor role
- */
-async function isEditorCheck(check: string): Promise<CheckResultError> {
-  try {
-    const user = config.bootData?.user;
-    if (!user) {
-      return {
-        requirement: check,
-        pass: false,
-        error: 'User information not available',
-        context: null,
-      };
-    }
-
-    // Editor or higher (Admin, Grafana Admin)
-    const isEditor = user.orgRole === 'Editor' || user.orgRole === 'Admin' || user.isGrafanaAdmin === true;
-
-    return {
-      requirement: check,
-      pass: isEditor,
-      error: isEditor ? undefined : `User role '${user.orgRole || 'none'}' does not have editor permissions`,
-      context: {
-        orgRole: user.orgRole,
-        isGrafanaAdmin: user.isGrafanaAdmin,
-        userId: user.id,
-      },
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Editor check failed: ${error}`,
-      context: { error },
-    };
-  }
-}
-
-/**
- * Data sources availability checking - verifies at least one data source exists
- *
- * Use cases:
- * - Tutorial prerequisites: ensure basic Grafana setup is complete
- * - Data-dependent features: verify data sources are available before querying
- * - Setup validation: check if initial configuration is done
- * - Onboarding flows: guide users through basic setup
- *
- * How it works:
- * - Counts total number of configured data sources
- * - Uses ContextService.fetchDataSources() API call
- * - Passes if any data sources exist (regardless of type or status)
- *
- * Example usage:
- * - data-requirements="has-datasources" - at least one data source exists
- * - Common prerequisite for data visualization tutorials
- * - Ensures users have completed basic Grafana setup
- *
- * What it checks:
- * - Total count > 0 (any data source type)
- * - Includes all configured data sources (enabled and disabled)
- * - Does not test connectivity (use datasource-configured for that)
- *
- * Difference from specific checks:
- * - has-datasources: any data source exists
- * - has-datasource:X: specific data source exists
- * - datasource-configured:X: specific data source exists and works
- */
-async function hasDatasourcesCheck(check: string): Promise<CheckResultError> {
-  try {
-    const dataSources = await ContextService.fetchDataSources();
-    return {
-      requirement: check,
-      pass: dataSources.length > 0,
-      error: dataSources.length > 0 ? undefined : 'No data sources found',
-      context: { count: dataSources.length, types: dataSources.map((ds) => ds.type) },
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Failed to check data sources: ${error}`,
-      context: { error },
-    };
-  }
-}
-
-/**
- * Plugin enabled status checking - verifies a specific plugin is installed AND enabled
- *
- * Use cases:
- * - Before using plugin features: ensure plugin is active
- * - Tutorial prerequisites: verify required plugins are enabled
- * - Feature availability: check if optional plugins are ready to use
- * - Troubleshooting: validate plugin activation status
- *
- * How it works:
- * - Finds plugin by exact ID match
- * - Checks both existence and enabled status
- * - Distinguishes between "not installed" vs "installed but disabled"
- *
- * Example usage:
- * - data-requirements="plugin-enabled:volkovlabs-rss-datasource" - RSS plugin ready
- * - data-requirements="plugin-enabled:grafana-clock-panel" - Clock panel enabled
- * - data-requirements="plugin-enabled:alexanderzobnin-zabbix-app" - Zabbix app active
- *
- * Difference from has-plugin:
- * - has-plugin: checks if plugin is installed (may be disabled)
- * - plugin-enabled: checks if plugin is installed AND enabled (ready to use)
- */
-async function pluginEnabledCheck(check: string): Promise<CheckResultError> {
-  try {
-    const pluginId = check.replace('plugin-enabled:', '');
-    const plugins = await ContextService.fetchPlugins();
-
-    // Find the specific plugin
-    const plugin = plugins.find((p) => p.id === pluginId);
-
-    if (!plugin) {
-      return {
-        requirement: check,
-        pass: false,
-        error: `Plugin '${pluginId}' not found`,
-        context: {
-          searched: pluginId,
-          totalPlugins: plugins.length,
-          suggestion: `Plugin '${pluginId}' is not installed. Install it first, then enable it.`,
-        },
-      };
-    }
-
-    const isEnabled = plugin.enabled;
-
-    return {
-      requirement: check,
-      pass: isEnabled,
-      error: isEnabled ? undefined : `Plugin '${pluginId}' is installed but not enabled`,
-      context: {
-        searched: pluginId,
-        pluginFound: true,
-        isEnabled: plugin.enabled,
-        suggestion: isEnabled
-          ? undefined
-          : `Plugin '${pluginId}' is installed but disabled. Enable it in Grafana plugin settings.`,
-      },
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Plugin enabled check failed: ${error}`,
-      context: { error },
-    };
-  }
-}
-
-/**
- * Dashboard existence checking - verifies at least one dashboard exists in the system
- *
- * Use cases:
- * - Tutorial prerequisites: ensure dashboards are available for examples
- * - Content validation: verify Grafana has dashboard content
- * - Setup completion: check if users have created any dashboards
- * - Feature availability: ensure dashboard features can be demonstrated
- *
- * How it works:
- * - Uses /api/search endpoint with type: 'dash-db'
- * - Checks for any non-deleted dashboards
- * - Limit 1 for efficiency (just need to know if any exist)
- *
- * Example usage:
- * - data-requirements="dashboard-exists" - at least one dashboard exists
- * - Useful before dashboard navigation tutorials
- * - Ensures dashboard features can be demonstrated
- *
- * What it checks:
- * - Any dashboard exists (regardless of name or content)
- * - Excludes deleted dashboards
- * - System-wide check across all organizations
- *
- * Difference from specific checks:
- * - dashboard-exists: any dashboard exists
- * - has-dashboard-named:X: specific dashboard by title exists
- *
- * Performance: Uses limit=1 for fast existence check
- */
-async function dashboardExistsCheck(check: string): Promise<CheckResultError> {
-  try {
-    const dashboards = await getBackendSrv().get('/api/search', {
-      type: 'dash-db',
-      limit: 1, // We just need to know if any exist
-      deleted: false,
-    });
-
-    const hasDashboards = dashboards && dashboards.length > 0;
-
-    return {
-      requirement: check,
-      pass: hasDashboards,
-      error: hasDashboards ? undefined : 'No dashboards found in the system',
-      context: {
-        dashboardCount: dashboards?.length || 0,
-      },
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Dashboard existence check failed: ${error}`,
-      context: { error },
-    };
-  }
-}
-
-/**
- * Data source configuration testing - verifies a specific data source is properly configured and working
- *
- * Use cases:
- * - Before querying data: ensure data source connection works
- * - Tutorial prerequisites: verify Prometheus/Loki/etc. is set up correctly
- * - Dashboard creation: check required data sources are functional
- * - Troubleshooting guides: validate data source connectivity
- *
- * How it works:
- * - Finds data source by name or type (case-insensitive)
- * - Calls Grafana's /api/datasources/uid/{uid}/test endpoint
- * - Only passes if test returns status: 'success'
- *
- * Example usage:
- * - data-requirements="datasource-configured:prometheus" - check Prometheus works
- * - data-requirements="datasource-configured:loki" - check Loki connection
- * - data-requirements="datasource-configured:My Custom DS" - check by exact name
- *
- * Difference from has-datasource:
- * - has-datasource: checks if data source exists/is installed
- * - datasource-configured: checks if data source exists AND connection test passes
- */
-async function datasourceConfiguredCheck(check: string): Promise<CheckResultError> {
-  try {
-    const dsRequirement = check.replace('datasource-configured:', '').toLowerCase();
-    const dataSources = await ContextService.fetchDataSources();
-
-    if (dataSources.length === 0) {
-      return {
-        requirement: check,
-        pass: false,
-        error: 'No data sources available to test',
-        context: {
-          searched: dsRequirement,
-          totalDataSources: 0,
-          suggestion: 'Configure at least one data source first',
-        },
-      };
-    }
-
-    // Find the specific data source to test
-    let targetDataSource = null;
-
-    // Check for exact matches in name or type, then normalized type (same logic as hasDataSourceCheck)
-    for (const ds of dataSources) {
-      if (ds.name.toLowerCase() === dsRequirement || ds.type.toLowerCase() === dsRequirement) {
-        targetDataSource = ds;
-        break;
-      }
-      const normalizedType = ds.type
-        .toLowerCase()
-        .replace(/^grafana-/, '')
-        .replace(/-datasource$/, '');
-      if (normalizedType === dsRequirement) {
-        targetDataSource = ds;
-        break;
-      }
-    }
-
-    if (!targetDataSource) {
-      return {
-        requirement: check,
-        pass: false,
-        error: `Data source '${dsRequirement}' not found`,
-        context: {
-          searched: dsRequirement,
-          totalDataSources: dataSources.length,
-          suggestion: `Data source '${dsRequirement}' not found. Check the name/type and ensure it exists.`,
-        },
-      };
-    }
-
-    try {
-      // Use the data source test API
-      const testResult = await getBackendSrv().post(`/api/datasources/uid/${targetDataSource.uid}/test`);
-
-      const isConfigured = testResult && testResult.status === 'success';
-
-      return {
-        requirement: check,
-        pass: isConfigured,
-        error: isConfigured
-          ? undefined
-          : `Data source '${targetDataSource.name}' test failed: ${testResult?.message || 'Unknown error'}`,
-        context: {
-          searched: dsRequirement,
-          testedDataSource: {
-            id: targetDataSource.id,
-            name: targetDataSource.name,
-            type: targetDataSource.type,
-          },
-          testResult: testResult?.status || 'unknown',
-          suggestion: isConfigured
-            ? undefined
-            : `Data source '${targetDataSource.name}' exists but configuration test failed. Check connection settings.`,
-        },
-      };
-    } catch (testError) {
-      // If test fails, it might still be configured but unreachable
-      return {
-        requirement: check,
-        pass: false,
-        error: `Data source configuration test failed: ${testError}`,
-        context: {
-          searched: dsRequirement,
-          testedDataSource: {
-            id: targetDataSource.id,
-            name: targetDataSource.name,
-            type: targetDataSource.type,
-          },
-          testError: String(testError),
-          suggestion: `Test API call failed for '${targetDataSource.name}'. Check data source permissions and connectivity.`,
-        },
-      };
-    }
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Data source configuration check failed: ${error}`,
-      context: { error },
-    };
-  }
-}
-
-/**
- * Guide response variable checking - verifies a stored response matches expected value
- *
- * Use cases:
- * - Gating sections based on user input (e.g., policy acceptance)
- * - Conditional content based on previous answers
- * - Workflow enforcement based on collected data
- * - Required field validation before proceeding
- *
- * How it works:
- * - Parses requirement format: var-{variableName}:{expectedValue}
- * - Looks up stored response from guideResponseStorage (async, syncs across devices)
- * - Supports wildcard (*) for any non-empty value
- * - Supports boolean string comparison (true/false)
- * - Supports exact string matching
- *
- * Example usage:
- * - var-policyAccepted:true - boolean must be true
- * - var-datasourceName:* - any non-empty value
- * - var-datasourceName:prometheus - exact match
- * - var-region:us-east-1 - exact string match
- *
- * Note: Currently uses a default guide ID. In a full implementation,
- * the guide ID should be passed through the check context.
- */
-async function guideVariableCheck(check: string): Promise<CheckResultError> {
-  try {
-    // Parse the requirement format: var-{variableName}:{expectedValue}
-    const match = check.match(/^var-([^:]+):(.+)$/);
-    if (!match) {
-      return {
-        requirement: check,
-        pass: false,
-        error: `Invalid variable requirement format: ${check}. Expected: var-{variableName}:{expectedValue}`,
-        context: { format: 'var-{variableName}:{expectedValue}' },
-      };
-    }
-
-    const variableName = match[1]!;
-    const expectedValue = match[2]!;
-
-    // Get current guide ID from URL or use default
-    // TODO: In a full implementation, pass guideId through context
-    const guideId = getCurrentGuideId();
-    const actualValue = await guideResponseStorage.getResponse(guideId, variableName);
-
-    // Check for wildcard (any non-empty value)
-    if (expectedValue === '*') {
-      const hasValue = actualValue !== undefined && actualValue !== '' && actualValue !== null;
-      return {
-        requirement: check,
-        pass: hasValue,
-        error: hasValue ? undefined : `Variable '${variableName}' has no value set`,
-        context: { variableName, expectedValue, actualValue, guideId },
-      };
-    }
-
-    // Check for boolean values
-    if (expectedValue === 'true') {
-      const isTrue = actualValue === true || actualValue === 'true';
-      return {
-        requirement: check,
-        pass: isTrue,
-        error: isTrue ? undefined : `Variable '${variableName}' is not true (got: ${actualValue})`,
-        context: { variableName, expectedValue, actualValue, guideId },
-      };
-    }
-
-    if (expectedValue === 'false') {
-      const isFalse = actualValue === false || actualValue === 'false';
-      return {
-        requirement: check,
-        pass: isFalse,
-        error: isFalse ? undefined : `Variable '${variableName}' is not false (got: ${actualValue})`,
-        context: { variableName, expectedValue, actualValue, guideId },
-      };
-    }
-
-    // Exact string match
-    const matches = String(actualValue) === expectedValue;
-    return {
-      requirement: check,
-      pass: matches,
-      error: matches ? undefined : `Variable '${variableName}' does not match '${expectedValue}' (got: ${actualValue})`,
-      context: { variableName, expectedValue, actualValue, guideId },
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Variable check failed: ${error}`,
-      context: { error: String(error) },
-    };
-  }
-}
-
-/**
- * Renderer context checking - verifies the current renderer context
- *
- * Use cases:
- * - Conditional content: show different content based on renderer context
- * - Context-specific features: enable features only in specific renderers
- * - Multi-renderer support: adapt behavior for different rendering contexts
- *
- * How it works:
- * - Checks renderer value from requirement string
- * - Returns true for renderer:pathfinder (always true in app context)
- * - Returns false for renderer:website (always false in app context)
- * - Other renderer values are treated as unknown and fail
- *
- * Example usage:
- * - data-requirements="renderer:pathfinder" - only in pathfinder app
- * - data-requirements="renderer:website" - only in website context (always false in app)
- *
- * Renderer values:
- * - pathfinder: Grafana Pathfinder app context (always true)
- * - website: Website/public documentation context (always false in app)
- *
- * Note: This requirement is evaluated differently by different tools to render
- * different content in different contexts.
- */
-async function rendererCheck(check: string): Promise<CheckResultError> {
-  try {
-    const rendererValue = check.replace('renderer:', '').toLowerCase();
-
-    if (rendererValue === 'pathfinder') {
-      return {
-        requirement: check,
-        pass: true,
-        error: undefined,
-        context: { renderer: rendererValue, context: 'app' },
-      };
-    }
-
-    if (rendererValue === 'website') {
-      return {
-        requirement: check,
-        pass: false,
-        error: `Renderer requirement '${check}' is not satisfied (website context is not available in app)`,
-        context: { renderer: rendererValue, context: 'app' },
-      };
-    }
-
-    return {
-      requirement: check,
-      pass: false,
-      error: `Unknown renderer value: '${rendererValue}'. Supported values: 'pathfinder', 'website'`,
-      context: { renderer: rendererValue, supported: ['pathfinder', 'website'] },
-    };
-  } catch (error) {
-    return {
-      requirement: check,
-      pass: false,
-      error: `Renderer check failed: ${error}`,
-      context: { error: String(error) },
-    };
-  }
-}
-
-/**
- * Check whether the Coda terminal is connected.
- * Reads from a module-level getter exposed by TerminalContext.
- */
-async function terminalActiveCheck(check: string): Promise<CheckResultError> {
-  try {
-    // Dynamic import to avoid circular dependencies and keep terminal
-    // code tree-shakeable when the feature is disabled
-    const { getTerminalConnectionStatus } = await import('../integrations/coda/TerminalContext');
-    const status = getTerminalConnectionStatus();
-    const isConnected = status === 'connected';
-    return {
-      requirement: check,
-      pass: isConnected,
-      error: isConnected ? undefined : 'Terminal is not connected. Connect to a terminal session to continue.',
-      context: { terminalStatus: status },
-    };
-  } catch {
-    return {
-      requirement: check,
-      pass: false,
-      error: 'Terminal integration is not available.',
-      context: null,
-    };
-  }
-}
-
-/**
- * Get the current guide ID from the global context set by ContentRenderer.
- * Falls back to 'default' if not available.
- */
-function getCurrentGuideId(): string {
-  try {
-    // Use the guide ID set by ContentRenderer
-    const guideId = (window as any).__DocsPluginGuideId;
-    if (guideId && typeof guideId === 'string') {
-      return guideId;
-    }
-    return 'default';
-  } catch {
-    return 'default';
-  }
-}
-
-/**
- * Validates interactive element props and logs errors for impossible configurations
- *
- * This utility detects impossible requirement combinations and logs console errors
- * to help authors catch configuration mistakes during development.
- *
- * Specifically, it checks if an element has 'exists-reftarget' requirement but
- * no refTarget, which would make the step impossible to pass.
- *
- * @param props - Interactive element props to validate
- * @param elementType - Human-readable element type (e.g., 'InteractiveStep', 'InteractiveMultiStep')
- * @returns true if validation passed, false if errors were logged
+ * Validates interactive element props and logs errors for impossible configurations.
+ *
+ * Specifically catches steps with `exists-reftarget` requirement but no refTarget,
+ * which would make the step impossible to pass. Used at author-time to surface
+ * mistakes in the block editor.
  */
 export function validateInteractiveRequirements(
   props: {
